+++ packages/server/src/request-handler.ts
@@ -11,7 +11,7 @@ import { HandlerBuilder, responseBuilder } from "./lib/handler";
 import { getMockSession, refreshToken } from "./mockAuth";
 import renderIndex, { renderFooter, renderHeader } from "./render-index";
 import { search } from "./search";
-import { getTaConfig } from "./task-analytics-service";
+import { getTaConfig } from "./task-analytics";
 import { texts } from "./texts";
 import UnleashService from "./unleash-service";
 import { validParams } from "./validateParams";
